requirements.txt
pydantic-ai-slim[mcp]==1.34.0
Flask==2.2.5
json_repair==0.51.0
Pillow==12.0.0
Requests==2.32.5
openai==2.9.0
volcengine-python-sdk[ark]